refactor: remove Claim page (#4200)
