--- apps/cowswap-frontend/src/legacy/state/claim/types.ts
@@ -1,31 +1,0 @@
-import { SyntheticEvent } from 'react'
-
-import { Currency, CurrencyAmount, Price, Token } from '@uniswap/sdk-core'
-
-import { UserClaimData } from './hooks'
-
-export type ClaimCommonTypes = {
-  account: string | null | undefined
-  hasClaims: boolean
-  isClaimed: boolean
-  claims: EnhancedUserClaimData[]
-  isAirdropOnly: boolean
-  tokenCurrencyAmount: CurrencyAmount<Token>
-  handleChangeAccount: (e: SyntheticEvent<HTMLButtonElement>) => void
-}
-
-// Enhanced UserClaimData with useful additional properties
-export type EnhancedUserClaimData = UserClaimData & {
-  claimAmount: CurrencyAmount<Token>
-  isFree: boolean
-  currencyAmount?: CurrencyAmount<Currency> | undefined
-  price?: Price<Currency, Currency> | undefined
-  cost?: CurrencyAmount<Currency> | undefined
-}
-
-export type InvestmentAmounts = {
-  vCowAmount?: CurrencyAmount<Currency>
-  investmentCost?: CurrencyAmount<Currency>
-}
-
-export type ClaimWithInvestmentData = EnhancedUserClaimData & InvestmentAmounts
