--- apps/cowswap-frontend/src/pages/Claim/ClaimNav.tsx
@@ -1,55 +1,0 @@
-import { shortenAddress } from '@cowprotocol/common-utils'
-import { ButtonSecondary } from '@cowprotocol/ui'
-import { Identicon } from '@cowprotocol/wallet'
-
-import CopyHelper from 'legacy/components/Copy'
-import { ClaimStatus } from 'legacy/state/claim/actions'
-import { useClaimDispatchers, useClaimState } from 'legacy/state/claim/hooks'
-import { ClaimCommonTypes } from 'legacy/state/claim/types'
-
-import { ClaimAccount, ClaimAccountButtons, TopNav } from './styled'
-
-type ClaimNavProps = Pick<ClaimCommonTypes, 'account' | 'handleChangeAccount'>
-
-export default function ClaimNav({ account, handleChangeAccount }: ClaimNavProps) {
-  const { activeClaimAccount, activeClaimAccountENS, claimStatus, investFlowStep } = useClaimState()
-  const { setActiveClaimAccount, setIsSearchUsed } = useClaimDispatchers()
-
-  const isDefaultStatus = claimStatus === ClaimStatus.DEFAULT
-  const isConfirmed = claimStatus === ClaimStatus.CONFIRMED
-  const hasActiveAccount = activeClaimAccount !== ''
-  const allowToChangeAccount = investFlowStep < 2 && (isDefaultStatus || isConfirmed)
-
-  return (
-    <TopNav>
-      <ClaimAccount>
-        <div>
-          {hasActiveAccount && (
-            <>
-              <Identicon account={activeClaimAccount} size={46} />
-              <p>{activeClaimAccountENS ? activeClaimAccountENS : shortenAddress(activeClaimAccount)} </p>
-              <CopyHelper toCopy={activeClaimAccount} />
-            </>
-          )}
-        </div>
-        <ClaimAccountButtons>
-          {allowToChangeAccount && hasActiveAccount ? (
-            <ButtonSecondary onClick={handleChangeAccount}>Change account</ButtonSecondary>
-          ) : (
-            !!account &&
-            allowToChangeAccount && (
-              <ButtonSecondary
-                onClick={() => {
-                  setActiveClaimAccount(account)
-                  setIsSearchUsed(false)
-                }}
-              >
-                Switch to connected account
-              </ButtonSecondary>
-            )
-          )}
-        </ClaimAccountButtons>
-      </ClaimAccount>
-    </TopNav>
-  )
-}
