--- apps/cowswap-frontend/src/pages/Claim/FooterNavButtons.tsx
@@ -1,139 +1,0 @@
-import { ReactNode, useEffect, useRef } from 'react'
-
-import { Command } from '@cowprotocol/types'
-import { ButtonPrimary, ButtonSecondary } from '@cowprotocol/ui'
-import { useWalletInfo } from '@cowprotocol/wallet'
-import { isAddress } from '@ethersproject/address'
-
-import { Trans } from '@lingui/macro'
-
-import { ClaimStatus } from 'legacy/state/claim/actions'
-import {
-  useClaimDispatchers,
-  useClaimState,
-  useHasClaimInvestmentFlowError,
-  useHasZeroInvested,
-} from 'legacy/state/claim/hooks'
-import { ClaimCommonTypes } from 'legacy/state/claim/types'
-
-import { ClaimAddressProps } from './ClaimAddress'
-import { FooterNavButtons as FooterNavButtonsWrapper } from './styled'
-
-type FooterNavButtonsProps = Pick<ClaimCommonTypes, 'hasClaims' | 'isClaimed' | 'isAirdropOnly'> &
-  Pick<ClaimAddressProps, 'toggleWalletModal'> & {
-    isPaidClaimsOnly: boolean
-    resolvedAddress: string | null
-    handleSubmitClaim: Command
-    handleCheckClaim: Command
-  }
-
-export default function FooterNavButtons({
-  hasClaims,
-  isClaimed,
-  isAirdropOnly,
-  isPaidClaimsOnly,
-  resolvedAddress,
-  toggleWalletModal,
-  handleSubmitClaim,
-  handleCheckClaim,
-}: FooterNavButtonsProps) {
-  const { account } = useWalletInfo()
-  const {
-    // account
-    activeClaimAccount,
-    // claiming
-    claimStatus,
-    // investment
-    investFlowStep,
-    isInvestFlowActive,
-    // table select change
-    selected,
-  } = useClaimState()
-
-  const {
-    // investing
-    setInvestFlowStep,
-    setIsInvestFlowActive,
-  } = useClaimDispatchers()
-
-  const hasError = useHasClaimInvestmentFlowError()
-  const hasZeroInvested = useHasZeroInvested()
-
-  const isInputAddressValid = isAddress(resolvedAddress || '')
-
-  const buttonRef = useRef<HTMLButtonElement>(null)
-
-  // User is connected and has some unclaimed claims
-  const isConnectedAndHasClaims = account && activeClaimAccount && hasClaims && claimStatus === ClaimStatus.DEFAULT
-  const noPaidClaimsSelected = !selected.length
-
-  useEffect(() => {
-    buttonRef?.current?.blur()
-  }, [activeClaimAccount])
-
-  let buttonContent: ReactNode = null
-
-  // Disconnected, show wallet connect
-  if (!account && activeClaimAccount && toggleWalletModal) {
-    buttonContent = (
-      <ButtonPrimary ref={buttonRef} onClick={toggleWalletModal}>
-        <Trans>Connect a wallet</Trans>
-      </ButtonPrimary>
-    )
-  }
-  // Already claimed
-  else if (isClaimed && claimStatus !== ClaimStatus.CONFIRMED && toggleWalletModal) {
-    buttonContent = (
-      <ButtonPrimary ref={buttonRef} onClick={toggleWalletModal} disabled>
-        <Trans>Already claimed</Trans>
-      </ButtonPrimary>
-    )
-  }
-  // User has no set active claim account and/or has claims, show claim account search
-  else if (!hasClaims && claimStatus === ClaimStatus.DEFAULT) {
-    buttonContent = (
-      <ButtonPrimary ref={buttonRef} disabled={!isInputAddressValid} type="text" onClick={handleCheckClaim}>
-        <Trans>Check claimable vCOW</Trans>
-      </ButtonPrimary>
-    )
-  }
-
-  // USER is CONNECTED + HAS SOMETHING TO CLAIM
-  if (isConnectedAndHasClaims) {
-    if (!isInvestFlowActive) {
-      buttonContent = (
-        <ButtonPrimary onClick={handleSubmitClaim} disabled={isPaidClaimsOnly && noPaidClaimsSelected}>
-          <Trans>Claim vCOW</Trans>
-        </ButtonPrimary>
-      )
-    } else if (!isAirdropOnly) {
-      buttonContent = (
-        <>
-          {investFlowStep === 0 ? (
-            <ButtonPrimary onClick={() => setInvestFlowStep(1)}>
-              <Trans>Continue</Trans>
-            </ButtonPrimary>
-          ) : investFlowStep === 1 ? (
-            <ButtonPrimary onClick={() => setInvestFlowStep(2)} disabled={hasError || hasZeroInvested}>
-              <Trans>Review</Trans>
-            </ButtonPrimary>
-          ) : (
-            <ButtonPrimary onClick={handleSubmitClaim}>
-              <Trans>Claim and invest vCOW</Trans>
-            </ButtonPrimary>
-          )}
-
-          <ButtonSecondary
-            onClick={() =>
-              investFlowStep === 0 ? setIsInvestFlowActive(false) : setInvestFlowStep(investFlowStep - 1)
-            }
-          >
-            <Trans>Go back</Trans>
-          </ButtonSecondary>
-        </>
-      )
-    }
-  }
-
-  return <FooterNavButtonsWrapper>{buttonContent}</FooterNavButtonsWrapper>
-}
