--- apps/cowswap-frontend/src/pages/Claim/const.ts
@@ -1,3 +1,0 @@
-import { isProd, isEns, isBarn } from '@cowprotocol/common-utils'
-
-export const IS_TESTING_ENV = !isProd && !isEns && !isBarn
